Refactor old accounting classes into DTOs. Deprecate old accounting CostEstimate and replace it with a DTO akin to the offer DTO. Update DTO description and tests
